refactor: extract different button sets into their own component
